Fix display contents by user condition branch

+++ app/javascript/packs/components/header.jsx
@@ -5,6 +5,7 @@ class Header extends Component {
   constructor(props){
     super(props);
     this.state = {
+      userResponse: false,
       isLoggedIn: false,
     }
     this.getCurrentUser  = this.getCurrentUser.bind(this)
@@ -17,27 +18,34 @@ class Header extends Component {
   getCurrentUser(){
     axios.get("/users/current")
     .then(res => {
-      if (res.data.id) {
-        this.setState({isLoggedIn: true})
+      if (res.data) {
+        this.setState({
+          userResponse: true,
+          isLoggedIn: true
+        })
       }else{
-        this.setState({isLoggedIn: false})
+        this.setState({
+          userResponse: true,
+          isLoggedIn: false
+        })
       }
     })
   }
 
   render() {
     let headerList;
-
-    if (this.state.isLoggedIn) {
-      headerList = <ul className="header__menu">
-                    <li className="header__list"><a href="/users/edit" className="header__link">設定</a></li>
-                    <li className="header__list"><a href="/users/sign_out" data-method="delete" className="header__link">ログアウト</a></li>
-                  </ul>
-    }else{
-      headerList = <ul className="header__menu">
-                    <li className="header__list"><a href="/users/sign_up" className="header__link">新規登録</a></li>
-                    <li className="header__list"><a href="/users/sign_in" className="header__link">ログイン</a></li>
-                  </ul>
+    if (this.state.userResponse) {
+      if (this.state.isLoggedIn) {
+        headerList = <ul className="header__menu">
+                      <li className="header__list"><a href="/users/edit" className="header__link">設定</a></li>
+                      <li className="header__list"><a href="/users/sign_out" data-method="delete" className="header__link">ログアウト</a></li>
+                    </ul>
+      }else{
+        headerList = <ul className="header__menu">
+                      <li className="header__list"><a href="/users/sign_up" className="header__link">新規登録</a></li>
+                      <li className="header__list"><a href="/users/sign_in" className="header__link">ログイン</a></li>
+                    </ul>
+      }
     }
 
     return (
